Move widget files to folders representing different levels of widgets
