+++ test/widgets/common/clickable_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:intellicook_mobile/widgets/common/clickable.dart';
+import 'package:intellicook_mobile/widgets/low_level/clickable.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
