+++ test/widgets/common/input_field_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:intellicook_mobile/widgets/common/input_field.dart';
+import 'package:intellicook_mobile/widgets/high_level/input_field.dart';
 
 import '../../fixtures.dart';
